+++ src/ts/Circle.ts
@@ -9,7 +9,8 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `toSVGString` function (deprecated). Use `drawutilssvg` instead.
- * @version  1.3.0
+ * @modified 2022-08-15 Added the `containsPoint` function.
+ * @version  1.4.0
  **/
 
 import { Line } from "./Line";
@@ -17,7 +18,7 @@ import { UIDGenerator } from "./UIDGenerator";
 import { Vector } from "./Vector";
 import { VertTuple } from "./VertTuple";
 import { Vertex } from "./Vertex";
-import { SVGSerializable, UID } from "./interfaces";
+import { SVGSerializable, UID, XYCoords } from "./interfaces";
 
 /**
  * @classdesc A simple circle: center point and radius.
@@ -80,6 +81,19 @@ export class Circle implements SVGSerializable {
     this.uid = UIDGenerator.next();
     this.center = center;
     this.radius = radius;
+  }
+
+  /**
+   * Check if the given circle is fully contained inside this circle.
+   *
+   * @method containsPoint
+   * @param {XYCoords} point - The point to check if it is contained in this circle.
+   * @instance
+   * @memberof Circle
+   * @return {boolean} `true` if the given point is inside this circle.
+   */
+  containsPoint(point: XYCoords): boolean {
+    return this.center.distance(point) < this.radius;
   }
 
   /**